+++ KnilKitTests/AASAURLSuggestorTests.swift
@@ -40,4 +40,49 @@ class AASAURLSuggestorTests: XCTestCase {
             XCTAssert(suggested == value, "Suggest for \(key) is \(suggested) but should be \(value).")
         }
     }
+
+    func testSuggestAASA() {
+        let hosts: [String] = [
+            "twitter.com",
+            "facebook.com",
+            "netflix.com",
+            "instagram.com",
+            "yelp.com",
+            "amazon.com",
+            "pinterest.com",
+            "www.opentable.com",
+            "flipboard.com",
+            "citymapper.com",
+            "foursquare.com",
+            "www.dropbox.com",
+            "www.airbnb.com",
+            "www.periscope.tv",
+            "www.google.com",
+            "medium.com",
+            "overcast.fm",
+            "www.swarmapp.com",
+            "flickr.com",
+            "www.shazam.com",
+            "www.imdb.com",
+            "www.nytimes.com",
+            "linkedin.com",
+            "open.spotify.com",
+            "soundcloud.com",
+            "www.patreon.com" // Redirected to /.well-known/
+        ]
+        let expectations = hosts.map { XCTestExpectation(description: $0) }
+
+        for (index, host) in hosts.enumerated() {
+            AASAURLSuggestor.suggestAASA(from: host) { (result) in
+                switch result {
+                case .value(_):
+                    expectations[index].fulfill()
+                case .error(let error):
+                    XCTFail(host + ". " + error.localizedDescription)
+                }
+            }
+        }
+
+        wait(for: expectations, timeout: 5.0)
+    }
 }