+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Duplicate Universal Link flow animation 
 
 ## [1.0.0(5)] - 2018-08-03
 ### Added