+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.0.0(6)] - 2018-08-05
 ### Added
 - Duplicate Universal Link flow animation
 - Handle sites that put AASA in /.well-known/